fix: some circle deps
