+++ packages/theme/src/common/color-tokens/activity-bar.ts
@@ -1,9 +1,9 @@
 import { localize } from '@opensumi/ide-core-common';
 
 import { Color } from '../../common/color';
-import { registerColor, transparent } from '../color-registry';
+import { registerColor } from '../color-registry';
+import { transparent } from '../utils';
 
-import { contrastBorder } from './base';
 import { hcActiveBorderColor } from './basic-color';
 import { SIDE_BAR_BORDER } from './sidebar';
