+++ packages/theme/src/common/color-tokens/breadcrumb.ts
@@ -1,6 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor, transparent, lighten, darken } from '../color-registry';
+import { registerColor } from '../color-registry';
+import { transparent, lighten, darken } from '../utils';
 
 import { foreground } from './base';
 import { editorBackground, editorWidgetBackground } from './editor';
