+++ packages/theme/src/common/color-tokens/custom/base.ts
@@ -1,6 +1,7 @@
 import { localize } from '@opensumi/ide-core-common';
 
-import { registerColor, transparent } from '../../color-registry';
+import { registerColor } from '../../color-registry';
+import { transparent } from '../../utils';
 import { iconForeground, foreground } from '../base';
 import { editorWarningForeground, editorErrorForeground, editorInfoForeground } from '../editor';
